# Godriver2259 Reduce memory consumption in Connection.CompressWireMessage().