--- a/x/mongo/driver/compression_test.go
+++ b/x/mongo/driver/compression_test.go
@@ -7,9 +7,12 @@
 package driver
 
 import (
+	"io"
+	"os"
 	"strconv"
 	"testing"
 
+	"github.com/klauspost/compress/zstd"
 	"github.com/stretchr/testify/assert"
 	"go.mongodb.org/mongo-driver/x/mongo/driver/wiremessage"
 )
@@ -37,6 +40,81 @@ func TestCompression(t *testing.T) {
 			decompressed, err := DecompressPayload(compressed, opts)
 			assert.NoError(t, err)
 			assert.Equal(t, payload, decompressed)
+		})
+	}
+}
+
+func TestZstdWindowSize(t *testing.T) {
+	tests := []struct {
+		inputSize  int
+		level      zstd.EncoderLevel
+		windowSize int
+	}{
+		{
+			inputSize:  512,
+			level:      zstd.EncoderLevelFromZstd(wiremessage.DefaultZstdLevel),
+			windowSize: 1024,
+		},
+		{
+			inputSize:  512000,
+			level:      zstd.EncoderLevelFromZstd(wiremessage.DefaultZstdLevel),
+			windowSize: 524288,
+		},
+		{
+			inputSize:  16000000,
+			level:      zstd.EncoderLevelFromZstd(wiremessage.DefaultZstdLevel),
+			windowSize: 16777216,
+		},
+		{
+			inputSize:  32000000,
+			level:      zstd.EncoderLevelFromZstd(wiremessage.DefaultZstdLevel),
+			windowSize: 16777216,
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(strconv.Itoa(test.inputSize), func(t *testing.T) {
+			windowSize := calcZstdWindowSize(test.inputSize, test.level)
+			assert.Equal(t, test.windowSize, windowSize)
+		})
+	}
+}
+
+func BenchmarkCompression(b *testing.B) {
+	payload := func() []byte {
+		f, err := os.Open("compression.go")
+		if err != nil {
+			panic(err)
+		}
+		buf, err := io.ReadAll(f)
+		if err != nil {
+			panic(err)
+		}
+		for i := 1; i < 10; i++ {
+			buf = append(buf, buf...)
+		}
+		return buf
+	}()
+
+	compressors := []wiremessage.CompressorID{
+		wiremessage.CompressorSnappy,
+		wiremessage.CompressorZLib,
+		wiremessage.CompressorZstd,
+	}
+
+	for _, compressor := range compressors {
+		b.Run(strconv.Itoa(int(compressor)), func(b *testing.B) {
+			opts := CompressionOpts{
+				Compressor: compressor,
+				ZlibLevel:  wiremessage.DefaultZlibLevel,
+				ZstdLevel:  wiremessage.DefaultZstdLevel,
+			}
+			for i := 0; i < b.N; i++ {
+				_, err := CompressPayload(payload, opts)
+				if err != nil {
+					b.Error(err)
+				}
+			}
 		})
 	}
 }