--- a/x/mongo/driver/topology/connection.go
+++ b/x/mongo/driver/topology/connection.go
@@ -638,10 +638,7 @@ func (c *Connection) CompressWireMessage(src, dst []byte) ([]byte, error) {
 		return dst, ErrConnectionClosed
 	}
 	if c.connection.compressor == wiremessage.CompressorNoOp {
-		if len(dst) == 0 {
-			return src, nil
-		}
-		return append(dst, src...), nil
+		return src, nil
 	}
 	_, reqid, respto, origcode, rem, ok := wiremessage.ReadHeader(src)
 	if !ok {