[GODRIVER-2259](https://jira.mongodb.org/browse/GODRIVER-2259)
[GODRIVER-2023](https://jira.mongodb.org/browse/GODRIVER-2023)

- Tune the Zstd compression.
- Use src slice directly when using no compression.