@qingyang-hu 2022-06-17
After the discussion with @matthewdale, I switched to [EncodeAll()](https://pkg.go.dev/github.com/klauspost/compress@v1.13.6/zstd#Encoder.EncodeAll) to reuse the encoder. Benchmarks show an even lower memory consumption in the concurrent scenario.