@qingyang-hu 2022-06-22
### Benchmark results from [BenchmarkCompressPayload](https://github.com/mongodb/mongo-go-driver/blob/5d7bd99b167dcb19b89fe617215eacd08393b0f3/x/mongo/driver/compression_test.go#L44):

- OLD:
> CompressorZstd-4                 249    4412763 ns/op 38611199 B/op      135 allocs/op

- NEW:
> CompressorZstd-4                   5166        265163 ns/op        2968 B/op           4 allocs/op

- DELTA with "-count=10" using [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat)
> name old time/op new time/op delta
CompressorZstd-4 4.38ms ± 1% 0.23ms ± 1% -94.72% (p=0.000 n=10+10)
name old alloc/op new alloc/op delta
CompressorZstd-4 38.6MB ± 0% 0.0MB ± 0% -99.99% (p=0.000 n=10+10)
name old allocs/op new allocs/op delta
CompressorZstd-4 135 ± 0% 4 ± 0% -97.04% (p=0.000 n=10+10)

### Benchmark results from [BenchmarkConnection](https://github.com/mongodb/mongo-go-driver/blob/5d7bd99b167dcb19b89fe617215eacd08393b0f3/x/mongo/driver/topology/connection_test.go#L931):

- OLD:
> BenchmarkConnection/CompressWireMessage_CompressorNoOp-4 3836367 260.9 ns/op 256 B/op 1 allocs/op

- NEW:
> BenchmarkConnection/CompressWireMessage_CompressorNoOp-4 17769693 68.02 ns/op 0 B/op 0 allocs/op

- DELTA with "-count=10" using [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat)
> name old time/op new time/op delta
Connection/CompressWireMessage_CompressorNoOp-4 261ns ± 3% 48ns ±43% -81.69% (p=0.000 n=10+10)
name old alloc/op new alloc/op delta
Connection/CompressWireMessage_CompressorNoOp-4 256B ± 0% 0B -100.00% (p=0.000 n=10+10)
name old allocs/op new allocs/op delta
Connection/CompressWireMessage_CompressorNoOp-4 1.00 ± 0% 0.00 -100.00% (p=0.000 n=10+10)